FEAT: Add more error reporting
When testing interactions, sometimes there are crashes (during ols) or
there is an inconsistent state (subset=0), and the program continues
with no error message or an unclear one.

Add additional error messages to more easily pinpoint the problem in the
code.
TODO: Don't allow the errors to occur in the first place. The subset
error should be fixed in a comming commit that guarantees all subsets
exist at least as many times as their containing superset. the ols error
and other errors will need additional work.