FIX: Create valid output if no interactions
If there are no interactions for a drug, a dummy pandas dataframe is
created. This doesn't work when combined with previous drugs that
contained interactions, and confuses downstream analysis of the output
file.

Ignore the repetitions requirement for interactions until the dataframe
has been created with the correct columns. Then, if the interaction
doesn't have enough repetitions, remove it from the dataframe, leaving
the columns behind.