FEAT: Update interaction identification method
Possible interactions are identified as consecutive nodes in a tree, and
are counted multiple times if they occur multiple times in a tree.
The consecutive nodes requirement means that for an interaction ABC to
be found, A, B, and C, must be present out of the mTry variables in
three consecutive nodes. This is very unlikely for lower mTry values. In
addition, allowing interactions to be counted multiple times in a tree
can lead to a single tree having an outsized effect, and such artifacts
as A>B>C>A causing the recognition of ABC, but not any of AB, AC or
BC, which causes downstream bugs.

Change the identification method for possible interactions so the
repetition value is the number of trees in which a potential interaction
occurs, and an interaction can be any set of features that exist in the
same branch of a tree, without the requirement that their nodes be
consecutive.