FIX: Comment out R test-set code
In the past, the workflow was used on data that was spilt into training
and test data. Currently the data is not set up in this way. The
test-set python handling was removed in commit 7a30130 and no longer
provides the files neded by the R script to calculate MSE and pearsonR
statistics.

Comment out the code in the Rscript that assumes there is test set data.